Setup: improve memory footprint of build-artifacts
This is supposed to improve the memory footprint when building artifacts. Building
artifacts is based on the ImplementationOfInterfaceFinder which uses \ReflectionClass.
As outlined in https://mantis.ilias.de/view.php?id=28848 using \ReflectionClass
requires a lot of memory in this scenario, as PHP never frees information about
classes. This patch dispatches the actual usage of \ReflectionClass to another
process to make memory usage smaller.