Services/Component: remove classes extending from ilService or ilModule
There are various classes extending from ilService or ilModule that are not used
anywhere but in one place. Most components, though, do not have an according class.
Although having a class as central entry point to a component seems to be a good
idea in general, the concept will probably be easier to design from scratch.